refactor(daemon): split server/tests.rs into per-domain test modules
The server/ split landed in PR #355 left tests.rs at 2338 LOC,
violating the no-source-file-over-1000-LOC rule that #355 was
supposed to enforce (and that #356/#357/#358 already enforced
elsewhere).

Split into server/tests/{pack,client_env,compiler_hash,cache_trim,
link_cache,fingerprint,server_ipc,pch,write_cached,post_link_hook}.rs
grouped by the server/ subject each test targets. Each new file is
well under 1K LOC (largest: server_ipc.rs at 667). Per-domain helpers
(test fixture builders, jobserver-env constructors, write_fake_linker,
start_daemon) live alongside the tests that use them; the only
cross-module helper (CacheDirEnvGuard, used by both link_cache and
server_ipc) lives in tests/mod.rs.

All 72 server::tests::* tests still pass (226 daemon lib unit tests
+ 10 ignored integration tests = 236 total, same as before); no
behavior change.
